separated  the new list into its component
