+++ frontend/src/components/FinancialNews/FinancialNews.js
@@ -1,0 +1,129 @@
+import "../../App.css";
+import { useState, useEffect } from "react";
+import { retrieveFinancialNews } from "../../util/apiClient";
+import Box from '@mui/material/Box';
+import Link from '@mui/material/Link';
+import Card from '@mui/material/Card';
+import CardContent from '@mui/material/CardContent';
+import CardMedia from '@mui/material/CardMedia';
+import Typography from '@mui/material/Typography';
+import Chip from '@mui/material/Chip';
+import Stack from '@mui/material/Stack';
+import Button from '@mui/material/Button';
+import InputTextField from "../InputTextField";
+import { Loading } from '../Loading';
+import useMediaQuery from '@mui/material/useMediaQuery';
+import FinancialNewsLists from "./FinancialNewsLists";
+
+export default function FinancialNews(props) {
+    const isMobileScreen = useMediaQuery('(max-width:414px)');
+    const { filter = false } = props;
+    const [newsLists, setNewsLists] = useState([]);
+    const [tempTopic, setTempTopic] = useState("");
+    const [newsTopic, setNewsTopic] = useState(["Stock", "Business", "Finance", "Bank", "Investment", "Trading", "Tesla", "Apple", "Facebook", "Cryptocurrency",]);
+
+    useEffect(() => {
+        async function fetchNewsLists() {
+            const newsRes = await retrieveFinancialNews(newsTopic);
+            setNewsLists(newsRes.data);
+        }
+        fetchNewsLists();
+    }, [newsTopic]
+    )
+
+    const handleAddNewsTopic = (e) => {
+        const updateNewsTopic = [...newsTopic];
+        updateNewsTopic.push(tempTopic);
+        setTempTopic("")
+        setNewsTopic(updateNewsTopic);
+    }
+
+    const handleInput = (e) => {
+        const newTempInput = e.value;
+        setTempTopic(newTempInput);
+    }
+
+    const handleDelete = (index) => {
+        newsTopic.splice(index, 1);
+        const updateNewsTopic = [...newsTopic];
+        setNewsTopic(updateNewsTopic);
+    }
+
+    return (
+        <div style={style.div}>
+            {newsLists.length > 0 ?
+                <>
+                    <div style={style.subDivHeading}>
+                        <Typography
+                            variant="h5"
+                            color="black"
+                            component="div"
+                            my={2}
+                        >
+                            {isMobileScreen ? "News" : "Financial News"}
+                        </Typography>
+                        {filter && <div style={{...style.subDivInputField.main, ...(isMobileScreen ? style.subDivInputField.sm : style.subDivInputField.lg)}}>
+                            <InputTextField updateVal={handleInput} inputFieldLabel={isMobileScreen ? "Categories" : "Input Categories"} size="small" displayInput={tempTopic} />
+                            <Button variant="contained" type="submit" onClick={handleAddNewsTopic} style={style.convertButton} >
+                                Add
+                            </Button>
+                        </div>}
+                    </div>
+                    {filter && <Stack direction="row" style={style.Stack}>
+                        {newsTopic?.map((topic, index) => (
+                            <Chip label={topic} variant="outlined" onDelete={() => handleDelete(index)} style={style.Chip} />
+                        ))}
+                    </Stack>}
+                    {newsLists.map(news => {
+                        return (
+                            <Link
+                                key={news.title}
+                                href={news.link}
+                                className="hoverCard"
+                                sx={sxStyle.Link}>
+                                {!isMobileScreen ? 
+                                    <Card sx={sxStyle.Card}>
+                                        {news.thumbnail !== null && !isMobileScreen && <CardMedia
+                                            component="img"
+                                            sx={sxStyle.CardMedia}
+                                            image={news.thumbnail}
+                                            alt="Live from space album cover"
+                                        />}
+                                        <Box sx={sxStyle.Box}>
+                                            <FinancialNewsLists news={news} />
+                                        </Box>
+                                    </Card> : 
+                                    <FinancialNewsLists news={news} />}
+                            </Link>
+                        )
+                    })
+                    }
+                </>
+                : <Loading />
+            }
+        </div>
+    )
+};
+
+const sxStyle = {
+    Box: { display: 'flex', flexDirection: 'column', width: "100%", flex: '1 0 auto', justifyContent: "space-between", 
+            width: "min-content", padding: "20px" },
+    Link: { width: '100%', textDecoration: "none", margin: "7px 0" },
+    Card: { display: 'flex', width: '100%' },
+    CardMedia: { width: 240, height: 180, objectFit: "cover" },
+}
+
+const style = {
+    div: { display: 'flex', flexDirection: "column", width: '100%' },
+    subDivHeading: { display: "flex", justifyContent: "space-between", alignItems: "center", marginBottom: "10px"},
+    subDivInputField: {
+        main: { display: "flex", alignItems: "center", justifyContent: "flex-end" },
+        lg: { width: "25%" },
+        sm: { width: "60%" }
+    },
+    Stack: { display: "flex", flexWrap: "wrap", margin: "0px 0px 12px 0px" },
+    Chip: { marginRight: "5px", marginBottom: "5px" },
+    convertButton: { marginLeft: "8px", marginTop: "-1.5px" },
+}
+
+
